requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.9
django-celery-beat==2.5.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
kombu==5.3.5
packaging==23.2
prompt-toolkit==3.0.43
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
